feat(test-runner-mocha): default to use spec reporter
